ai-service/requirements.txt
flask
deepface
opencv-python
pillow
numpy